+++ tests/sync/util.js
@@ -125,6 +125,8 @@ const syncFromDisk = require('../../sync/syncFromDisk');
  * @property {string} set
  * @property {string} number
  * @property {boolean} allowIssueReporting
+ * @property {boolean} allowRealTimeGrading
+ * @property {boolean} requireHonorCode
  * @property {boolean} multipleInstance
  * @property {boolean} shuffleQuestions
  * @property {AssessmentAllowAccess[]} allowAccess