+++ ChangeLog.md
@@ -70,6 +70,8 @@
   * Add support for Rust to `pl-code` element (Nathan Walters).
 
   * Add tests for LTI auth (Dave Mussulman).
+  
+  * Add more robust Python autograder to `prairielearn/grader-python` (Nathan Bowman and Nicolas Nytko).
 
   * Add choose course instance dropdown to instructor nav-bar when viewing course only (Tim Bretl).
 
@@ -78,6 +80,22 @@
   * Add explanation of `None` role in documentation and instructor gradebook view (James Balamuta).
 
   * Add support for editing from bare git repo (Tim Bretl).
+  
+  * Add instructor panel to `course` question pages (Tim Yang).
+  
+  * Add `disregard-extra-elements` attribute to `pl-drawing` element to ignore duplicate grading objects (Nicolas Nytko).
+  
+  * Add extra `/course*` mount points (Tim Yang).
+
+  * Add R vector/matrix support in `pl-variable-output` (James Balamuta).
+
+  * Add example course question templates (James Balamuta).
+
+  * Add extra `/course*` mount points (Tim Yang).
+  
+  * Add confirmation dialog for unsaved changes on question pages (Tim Yang).
+
+  * Add Python library regex to plbase (Nathan Bowman).
 
   * Change v3 questions to disable autocomplete on the question form (Nathan Walters).
 
@@ -122,6 +140,8 @@
   * Change order in which course instances are listed to be by earliest and latest access date (Tim Bretl).
 
   * Change size limit for form data to account for increase from encoding (Nathan Bowman).
+
+  * Change styling for invalid input and `pl-string-input` for increased clarity. (Nicolas Nytko)
 
   * Fix dead letter cron job for `async` v3 (Matt West).
 
@@ -229,7 +249,11 @@
 
   * Fix gradebook and question statistics download links (Tim Bretl).
 
+  * Fix `demoRandomPlot` by updating the matlibplot subplot code (James Balamuta).
+
   * Remove `number` column from `course_instances` table and `number` property from `infoCourseInstance.json` schema (Tim Bretl).
+  
+  * Remove introduction alert at the top of `homework` assessments (Tim Yang).
 
 * __3.2.0__ - 2019-08-05